Added spawner and other stuff

+++ global/anchor.go
@@ -12,7 +12,7 @@ const (
 	AnchorLeft
 )
 
-func AnchorPosition(anchor int, xOffset, yOffset float32) (float32, float32) {
+func AnchorPosition(anchor int, xOffset, yOffset float64) (float64, float64) {
 	switch anchor {
 	case AnchorNone:
 		return xOffset, yOffset
@@ -21,15 +21,15 @@ func AnchorPosition(anchor int, xOffset, yOffset float32) (float32, float32) {
 	case AnchorTop:
 		return xOffset, 0
 	case AnchorTopRight:
-		return float32(ScreenWidth()) + xOffset, yOffset
+		return float64(ScreenWidth()) + xOffset, yOffset
 	case AnchorRight:
-		return float32(ScreenWidth()) + xOffset, yOffset
+		return float64(ScreenWidth()) + xOffset, yOffset
 	case AnchorBottomRight:
-		return float32(ScreenWidth()) + xOffset, float32(ScreenHeight()) + yOffset
+		return float64(ScreenWidth()) + xOffset, float64(ScreenHeight()) + yOffset
 	case AnchorBottom:
-		return xOffset, float32(ScreenHeight()) + yOffset
+		return xOffset, float64(ScreenHeight()) + yOffset
 	case AnchorBottomLeft:
-		return xOffset, float32(ScreenHeight()) + yOffset
+		return xOffset, float64(ScreenHeight()) + yOffset
 	}
 	return xOffset, yOffset
 }
