+++ global/distance.go
@@ -2,8 +2,8 @@ package global
 
 import "math"
 
-func Distance(x1, y1, x2, y2 float32) float32 {
+func Distance(x1, y1, x2, y2 float64) float64 {
 	dx := x1 - x2
 	dy := y1 - y2
-	return float32(math.Sqrt(float64(dx*dx + dy*dy)))
+	return math.Sqrt(float64(dx*dx + dy*dy))
 }
